работа с остновной таблицей и js

+++ finder/models.py
@@ -11,7 +11,7 @@ class Remains(models.Model):
     base_unit = models.CharField(max_length=10, null=True, verbose_name='Единица')
     project = models.CharField(max_length=30, null=True, verbose_name='Проект')
     quantity = models.FloatField(blank=True, null=True, verbose_name='Количество')
-    price = models.IntegerField(null=True)
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=True)  # Используйте DecimalField
 
 
     def __str__(self):
